+++ src/main/java/lotto/controller/LottoMain.java
@@ -1,10 +1,10 @@
 package lotto.controller;
 
-import lotto.domain.LottoComparator;
-import lotto.domain.LottoMachine;
-import lotto.domain.PurchasedLottoTickets;
-import lotto.domain.UserPurchase;
-import lotto.domain.WinningLottoNumbers;
+import lotto.domain.ticketresult.LottoComparator;
+import lotto.domain.ticketgenerator.LottoMachine;
+import lotto.domain.ticketpurchase.PurchasedLottoTickets;
+import lotto.domain.ticketpurchase.UserPurchase;
+import lotto.domain.ticketresult.WinningLottoNumbers;
 import lotto.view.InputView;
 import lotto.view.OutputView;
