+++ src/main/java/lotto/type/LottoMatchType.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.type;
 
 import java.util.Arrays;
 import java.util.List;
