+++ src/main/java/lotto/view/OutputView.java
@@ -2,9 +2,9 @@ package lotto.view;
 
 import java.util.List;
 import java.util.Map;
-import lotto.domain.LottoMatchType;
+import lotto.type.LottoMatchType;
 import lotto.domain.LottoTicket;
-import lotto.domain.PurchasedLottoTickets;
+import lotto.domain.ticketpurchase.PurchasedLottoTickets;
 import lotto.view.printer.InputPrinter;
 import lotto.view.printer.OutputPrinter;
