+++ src/test/java/lotto/domain/ticketgenerator/LottoGeneratorTest.java
@@ -1,6 +1,8 @@
-package lotto.domain;
+package lotto.domain.ticketgenerator;
 
 import java.util.Arrays;
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoTicket;
 
 public class LottoGeneratorTest {
     private final LottoTicket lottoTicket = new LottoTicket(
