+++ todo.md
@@ -1,6 +1,4 @@
 # todo
-- 패키지 정리 (메인 패키지, 테스트 패키지)
-  
 - 클래스다이어그램 필드, 메소드 동기화
 - 클래스다이어그램 의존성 확인.
 
@@ -20,3 +18,4 @@
 - LottoMatchType 관련 함수 LottoMatchType 클래스 내부로 이동.
 - PurchasedLottoTickets size() 디미터 법칙 적용. (OutputView.printPurchasedLottoTickets()).
 - LottoMain 각 함수 10줄 미만 리팩토링.
+- 패키지 정리 (메인 패키지, 테스트 패키지).
